Move action and utility functions to appropriate modules
