+++ bin/gftools-fix-cmap.py
@@ -17,76 +17,10 @@
 from os.path import basename
 from argparse import ArgumentParser
 from fontTools.ttLib import TTFont
-from fontTools.ttLib.tables._c_m_a_p import CmapSubtable
-
+from gftools.fix import convert_cmap_subtables_to_v4, drop_nonpid0_cmap, drop_mac_cmap
 
 description = "Manipulate a collection of fonts' cmap tables."
 
-
-def convert_cmap_subtables_to_v4(font):
-  """Converts all cmap subtables to format 4.
-
-  Returns a list of tuples (format, platformID, platEncID) of the tables
-  which needed conversion."""
-  cmap = font['cmap']
-  outtables = []
-  converted = []
-  for table in cmap.tables:
-    if table.format != 4:
-      converted.append((table.format, table.platformID, table.platEncID))
-    newtable = CmapSubtable.newSubtable(4)
-    newtable.platformID = table.platformID
-    newtable.platEncID = table.platEncID
-    newtable.language = table.language
-    newtable.cmap = table.cmap
-    outtables.append(newtable)
-  font['cmap'].tables = outtables
-  return converted
-
-
-def partition_cmap(font, test, report=True):
-  """Drops all cmap tables from the font which do not pass the supplied test.
-
-  Arguments:
-    font: A ``TTFont`` instance
-    test: A function which takes a cmap table and returns True if it should
-      be kept or False if it should be removed from the font.
-    report: Reports to stdout which tables were dropped and which were kept.
-
-  Returns two lists: a list of `fontTools.ttLib.tables._c_m_a_p.*` objects
-  which were kept in the font, and a list of those which were removed."""
-  keep = []
-  drop = []
-
-  for index, table in enumerate(font['cmap'].tables):
-    if test(table):
-      keep.append(table)
-    else:
-      drop.append(table)
-
-  if report:
-    for table in keep:
-        print(("Keeping format {} cmap subtable with Platform ID = {}"
-               " and Encoding ID = {}").format(table.format,
-                                               table.platformID,
-                                               table.platEncID))
-    for table in drop:
-        print(("--- Removed format {} cmap subtable with Platform ID = {}"
-             " and Encoding ID = {} ---").format(table.format,
-                                                 table.platformID,
-                                                 table.platEncID))
-
-  font['cmap'].tables = keep
-  return keep, drop
-
-
-def drop_nonpid0_cmap(font, report=True):
-  keep, drop = partition_cmap(font, lambda table: table.platformID == 0, report)
-  return drop
-
-def drop_mac_cmap(font, report=True):
-  keep, drop = partition_cmap(font, lambda table: table.platformID != 1 or table.platEncID != 0, report)
-  return drop
 
 def main():
   parser = ArgumentParser(description=description)
